Remove the Bugs And Community Page UI
