+++ Frontend/package.json
@@ -13,11 +13,13 @@
     "@chakra-ui/react": "^3.7.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
+    "@heroicons/react": "^2.2.0",
     "@iconify/react": "^5.2.0",
     "@mui/icons-material": "^6.4.4",
     "@mui/joy": "^5.0.0-beta.51",
     "@mui/material": "^6.4.3",
     "@mui/x-date-pickers": "^7.25.0",
+    "@tailwindcss/vite": "^4.0.7",
     "axios": "^1.7.9",
     "dayjs": "^1.11.13",
     "framer-motion": "^12.4.0",
