+++ Frontend/src/components/tester/CommunityDetails.jsx
