+++ Frontend/src/components/tester/styles/CommunityDetails.css
