+++ Frontend/src/components/tester/styles/CommunityList.css
