+++ Frontend/src/components/tester/styles/CreateCommunity.css
